refactor member api helpers into their own file
