+++ packages/cli/src/pages/api/lists/[id]/__integration__/members.test.ts
@@ -1,9 +1,9 @@
 import { apiLogin } from "../../../__integration__/util/login";
+import { apiCreateList } from "../../../__integration__/util/lists";
 import {
-  apiCreateList,
   apiGetListMembers,
   apiCreateListMember,
-} from "../../../__integration__/util/lists";
+} from "../../../__integration__/util/listMember";
 
 describe("lists/[id]/members", () => {
   describe("not logged in", () => {
